% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visaNodeLinkChart.R
\name{visaNodeLinkChart}
\alias{visaNodeLinkChart}
\title{visa charts 5.0.5}
\usage{
visaNodeLinkChart(
  tagName,
  linkData,
  nodeData,
  propList,
  width = NULL,
  height = NULL,
  ...
)
}
\arguments{
\item{tagName}{String. The custom web component HTML tag for the Visa Chart Component. Set by respective chart functions.}

\item{linkData}{a valid R data frame. See more details in respective component functions.}

\item{nodeData}{a valid R data frame. See more details in respective component functions.}

\item{propList}{a list of props, created by each component function, see \href{https://github.com/visa/visa-chart-components}{Visa Chart Components}.}

\item{width}{Number. Width of chart container.}

\item{height}{Number. Height of chart container.}

\item{...}{All other props passed into the function will be passed through to the chart.}
}
\value{
a visaNodeLinkChart htmlwidget object for creating a variety of plot types
}
\description{
Visa Chart Components wrapped in r htmlwidgets package
}
